backend/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
curl_cffi==0.6.1
Django==4.2.10
django-cors-headers==4.3.1
duckduckgo_search==4.5.0
idna==3.6
lxml==5.1.0
mysqlclient==2.2.3
pycparser==2.21
python-dotenv==1.0.1
RateMyProfessorAPI==1.3.6
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
